# atodo-api-spec.yaml
...
openapi: 3.0.3
info:
  title: A-To-Do API
  description: >-
    Backend API for the A-To-Do to-do list application. Implementing this
    spec turns the app from a single-browser, localStorage-only prototype
    into a real multi-tenant service: every endpoint below is scoped to the
    authenticated account (via its bearer token), so two people logging into
    two different accounts genuinely see two different task lists, rather
    than the one shared list the pre-backend prototype used.

    All endpoints are prefixed with `/atodo/v1`. The bearer token is an
    opaque, server-issued credential (a real signed JWT is the intended
    implementation) that embeds at least `sub` (the account id) and the
    account's current `subscription` claim, so the client can read
    subscription status without an extra round trip -- see the `Token`
    schema. Whenever an endpoint changes something the token's claims
    reflect (starting/cancelling/scheduling deletion of a subscription), it
    returns a freshly issued token in place of the old one; the client is
    expected to discard the old one immediately.

    Every error response uses the same `Error` schema (`{ code, message }`).
    `code` is a stable, machine-readable identifier the client branches on
    (e.g. to show a specific translated message); `message` is a
    human-readable fallback, not meant to be shown as-is in a finished UI.
  version: "1.0.0"
...
servers:
  - url: "{baseUrl}/atodo/v1"
    variables:
...
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    patch:
      tags: [Users]
      summary: Update the current account's profile/preferences
      description: >-
        Partial update -- send only the fields you're changing. Covers
        everything Settings can change (nickname, avatar, time format,
        background, language) plus the small bits of session state the
        client persists across reloads (which task is currently
        active/focused, and the to-do list's current view mode). None of
        this affects the bearer token's own claims, so the token stays
        valid and unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname: { type: string }
                avatar:
                  type: string
                  nullable: true
                  description: A data URL (the client square-crops/downscales client-side before sending) or null to remove it.
                timeFormat: { type: string, enum: ["12", "24"] }
                background:
                  allOf: [{ $ref: "#/components/schemas/Background" }]
                  nullable: true
                language:
                  type: string
                  enum: [en, hr]
                  nullable: true
                activeTaskId:
                  type: string
                  nullable: true
                activeOccurrenceDate:
                  type: string
                  format: date
                  nullable: true
                todoViewMode:
                  type: string
                  enum: [pending, next-recurrence, all]
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [Users]
      summary: Permanently delete the current account and all its data, immediately
      description: >-
        Deletes the account row and every task/note/preference tied to it.
        Takes effect immediately and can't be undone -- this is the
        "Delete immediately" choice in Settings' danger zone. A paying
        subscriber choosing to keep access until their current billing
        period ends instead wants `POST /users/me/schedule-deletion`, not
        this.
      responses:
        "204":
          description: Deleted. The bearer token used for this request is invalid from this point on.
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /users/me/schedule-deletion:
    post:
      tags: [Users]
      summary: Schedule the account for deletion once its current subscription period ends
      description: >-
        The alternative to `DELETE /users/me` for a paying subscriber who
        doesn't want to forfeit the rest of a period they already paid for:
        keeps full access until `subscription.expiresAt`, but immediately
        sets `subscription.cancelAtPeriodEnd` (there's nothing left to renew
        into) and `subscription.scheduledDeletion`. The account is actually
        deleted, server-side, once `expiresAt` passes -- via whatever
        mechanism enforces that (a scheduled job, or lazily the next time
        `POST /auth/login`/`GET /auth/me` runs past that point, matching
        this app's existing inactivity-deletion enforcement) -- unless
        `DELETE /users/me/schedule-deletion` cancels it first.
      responses:
        "200":
          description: Scheduled.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { $ref: "#/components/schemas/Token" }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: No active subscription to schedule deletion against.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: NO_ACTIVE_SUBSCRIPTION, message: "There's no active subscription to schedule deletion for." }
    delete:
      tags: [Users]
      summary: Cancel a previously scheduled deletion
      description: >-
        Undoes `POST /users/me/schedule-deletion`'s `scheduledDeletion`
        flag. Deliberately leaves `cancelAtPeriodEnd` alone: "cancel the
        deletion" only promises to keep the account around, not to
        silently resume billing the account holder never explicitly asked
        to resume.
      responses:
        "200":
          description: Cancelled.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { $ref: "#/components/schemas/Token" }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Missing/invalid bearer token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: No subscription (scheduled for deletion or otherwise) to act on.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: NO_SUBSCRIPTION, message: "There's no subscription to cancel deletion for." }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <token>`, as returned by `POST /auth/login`
        or any of the subscription-mutating endpoints. See the `Token`
        schema for what it's expected to carry.
  schemas:
    EpochMillis:
      type: integer
      format: int64
      description: >-
        Milliseconds since the Unix epoch (`Date.now()`/`new Date(ms)` on the
        client), not an RFC 3339 string -- deliberately, since every one of
        these fields is used in client-side arithmetic/sorting (elapsed
        time, "is this still active", "which of these was created first",
        ...) and epoch millis need no parsing to do that. `dueDate`/
        `endDate`/occurrence-date fields elsewhere are a different thing --
        plain `YYYY-MM-DD` strings with no time component, compared as
        strings, so those stay `format: date`.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable -- the client branches on this, not `message`.
        message:
          type: string
          description: Human-readable fallback; a finished client should show its own translated copy per `code` instead.
    Token:
      type: string
      description: >-
        Opaque to the client beyond reading its claims (a real
        implementation: a signed JWT). Expected claims: `sub` (the account
        id) and `subscription` (the same shape as `User.subscription`, a
        snapshot as of when the token was issued -- re-issued by every
        endpoint that changes it, see each one's own description). The
        client never trusts a locally decoded token's claims for anything
        but quick, non-authoritative UI reads (e.g. "should the Subscribe
        button show right now") -- `GET /auth/me` is the actual source of
        truth for whether the account/subscription are still genuinely in
        that state.
    Background:
      type: object
      nullable: true
      description: The Settings background picker's currently-applied Unsplash photo, or null for none.
      properties:
        regularUrl: { type: string, format: uri }
        thumbUrl: { type: string, format: uri }
        photographerName: { type: string }
        photographerUrl: { type: string, format: uri }
        photoLink: { type: string, format: uri }
    Subscription:
      type: object
      nullable: true
      description: null means the account has never subscribed (the "Free" plan).
      properties:
        id: { type: string }
        plan:
          type: string
          enum: [trial, pro]
        billingInterval:
          type: string
          enum: [monthly, annual]
          nullable: true
          description: null for a trial.
        startedAt: { $ref: "#/components/schemas/EpochMillis" }
        expiresAt:
          allOf: [{ $ref: "#/components/schemas/EpochMillis" }]
          description: >-
            The single source of truth for whether the subscription is
            currently active -- always `now < expiresAt`, never a separately
            cached boolean, so an expiry is correctly detected the instant
            it passes rather than only once something happens to re-check it.
        cancelAtPeriodEnd:
          type: boolean
          description: True once the subscription won't renew (see POST /subscriptions/cancel and POST /users/me/schedule-deletion) -- access is untouched until expiresAt regardless.
        scheduledDeletion:
          type: boolean
          description: True if the account itself, not just the subscription, is set to be deleted once expiresAt passes -- see POST /users/me/schedule-deletion.
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        nickname: { type: string }
        avatar:
          type: string
          nullable: true
          description: A data URL, or null for the initials-only fallback avatar.
        timeFormat: { type: string, enum: ["12", "24"] }
        background: { $ref: "#/components/schemas/Background" }
        language:
          type: string
          enum: [en, hr]
          nullable: true
          description: null means never chosen or IP-detected yet -- the client detects and PATCHes a real value the first time this is null right after login.
        activeTaskId:
          type: string
          nullable: true
          description: Which task (by Task.id) is currently focused/being timed, if any.
        activeOccurrenceDate:
          type: string
          format: date
          nullable: true
          description: Which occurrence of activeTaskId, for a recurring task that can show more than one row at once. Always null when activeTaskId is null.
        todoViewMode:
          type: string
          enum: [pending, next-recurrence, all]
        subscription: { $ref: "#/components/schemas/Subscription" }
...